% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAT.R
\name{BEAT}
\alias{BEAT}
\title{BEAT design}
\usage{
BEAT(K, Kstar, t, nk, truebio, deltaeta, probfunc)
}
\arguments{
\item{K}{Number of blocks}

\item{Kstar}{After Kstar blocks' enrollment, if estimate of biomarker is still 1, stop for futility.}

\item{t}{User-specified parameter that is used to flexibly define the enrollment cutoff.}

\item{nk}{A vector of length K with number of enrollment per arm per block}

\item{truebio}{True biomarker value}

\item{deltaeta}{A vector containing four values: deltapos, etapos, deltaneg, etaneg for PPoS and PPoF}

\item{probfunc}{A vector of length 2 response probablity functions for control and treatment arms}
}
\value{
(1) biomarker estimator and estimated standard deviation;
(2) stop indicator, screened number;
(3) enrolled and excluded positive number;
(4) treatment effect for positive, negative and overall (rejection probability, estimator, estimated standard deviation, 95\% cover probability based on estimated / true biomarker threshold)
}
\description{
Conduct BEAT design and obtain all important design characteristics
}
\examples{
K=4; Kstar=3; t=0.5; nk=c(111,111,110,110); truebio=0.6; deltaeta=c(0.5,0.4,0.7,0.7);
probfunc=c('exp(-2.2)/(1+exp(-2.2))', 'exp(-4+3*x)/(1+exp(-4+3*x))')
BEAT(K, Kstar, t, nk, truebio, deltaeta, probfunc)

}
