\encoding{latin1}
\name{BEAT}
\alias{BEAT-package}
\alias{BEAT}
\docType{package}
\title{
BEAT - BS-Seq Epimutation Analysis Toolkit
}
\description{
Modelling, data preparation and analysis of BS-Seq derived, region-based epimutation data
}
\details{
\tabular{ll}{
Package: \tab BEAT - Rpackage\cr
Type: \tab Package\cr
Version: \tab 0.99.1\cr
Date: \tab 2013-01-10\cr
License: \tab LGPL version 3 or later\cr
Maintainer: \tab Kemal Akman, akmank@mpipz.mpg.de\cr
}
}
\keyword{package}
