% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat.R
\name{beat}
\alias{beat}
\title{Run Beat}
\usage{
beat(input_counts, input_metadata, output_dir, dataset_name, original)
}
\arguments{
\item{gene_counts}{the path to the gene counts with rows as genes and columns as samples}
}
\description{
Takes in gene counts, metadata, and output paths, as well as a dataset name and outputs an html report.
}
