% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beat.R
\name{multi_beat}
\alias{multi_beat}
\title{Run MultiBeat}
\usage{
multi_beat(parent_dir, output_dir, output_name)
}
\arguments{
\item{parent_dir}{the path to the gene counts with rows as genes and columns as samples}

\item{output_dir}{the path to the directory in which the output shall be.}

\item{output_name}{the name of the report that will be generated.}
}
\description{
Takes in gene counts, metadata, and output paths, as well as a dataset name and outputs an html report.
}
